First commit.
This is a copy of my project 'LocalTemperature' which is tored on my mac- bbok.
Reason I moved it here is that Android Studio did not accept the original
files. For reasons I don't know, after the pull the manifest file was missing.

So I decided to abandon the project on my mac, delete it's repo and move
everything on my linux machine....